+++ config/custom-environment-variables.json
@@ -55,5 +55,8 @@
   },
   "rabbitMQ": {
     "url": "CLOUDAMQP_URL"
+  },
+  "kadepay": {
+    "uri": "KADEPAY_URI"
   }
 }
