+++ src/routes/transaction.js
@@ -6,10 +6,6 @@ const {AccessControl} = require('../interceptors');
 
 const router = express.Router({});
 
-router.post('/load-kadima', transactionController.topupConsumerWallet);
-
-router.get('/hello', transactionController.hello);
-
 // endpoint to send email for a paid cash transaction
 router.post('/:transactionId/receipt/:receiptId/email', transactionController.emailReceipt);
